Refactor out adjudicatingAcceptRoute
